backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
langchain
langchain-upstage
pillow==10.4.0
pdf2image==1.17.0
python-dotenv==1.0.1
pydantic==2.8.2
